+++ app/src/main/java/woowacourse/shopping/data/entity/RecentlyViewedProductEntity.kt
@@ -5,6 +5,6 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "RecentlyViewedProducts")
 class RecentlyViewedProductEntity(
-    @PrimaryKey val productUid: Int,
+    @PrimaryKey val productUid: Long,
     val timestamp: Long = System.currentTimeMillis(),
 )
